requirements.txt

streamlit==1.38.0
pandas==2.2.2
openpyxl==3.1.5
pdfplumber==0.11.4
PyPDF2==3.0.1
pdfminer.six==20231228
